refactor: auth 관련 파일들 index 처리
