+++ src/widgets/RegisterUser/index.ts
@@ -1,0 +1,2 @@
+export { PortfolioStep } from './PortfolioStep';
+export { ProfileStep } from './ProfileStep';
